backend/pyproject.toml
[project]
name = "pdf-accessibility-backend"
version = "0.1.0"
description = "PDF accessibility remediation pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.7.0",
    "pikepdf>=9.4.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.2.0",
    "Pillow>=11.0.0",
    "ocrmypdf>=16.7.0",
    "docling>=2.70.0",
    "docling-parse>=5.5.0",
    "fonttools>=4.0.0",
    "ftfy>=6.3.1",
    "rapidfuzz>=3.14.0",
    "regex>=2025.0.0",
    "scipy>=1.17.0",
    "rtree>=1.4.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"app/services/document_intelligence_models.py" = ["E741"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
